fix: 스키마 필드 타입 수정

+++ prisma/migrations/20251020150641_init/migration.sql
@@ -1,7 +1,7 @@
 -- CreateTable
 CREATE TABLE "analytics" (
     "analytics_id" BIGSERIAL NOT NULL,
-    "child_id" BIGINT NOT NULL,
+    "child_id" INT NOT NULL,
     "conversation_id" BIGINT NOT NULL,
     "extracted_keywords" JSONB NOT NULL,
     "created_at" TIMESTAMP(3) NOT NULL DEFAULT CURRENT_TIMESTAMP,
@@ -12,7 +12,7 @@ CREATE TABLE "analytics" (
 -- CreateTable
 CREATE TABLE "child_interests" (
     "interest_id" BIGSERIAL NOT NULL,
-    "child_id" BIGINT NOT NULL,
+    "child_id" INT NOT NULL,
     "keyword" VARCHAR(100) NOT NULL,
     "raw_score" DOUBLE PRECISION NOT NULL,
     "last_updated" TIMESTAMP(3) NOT NULL,
